recommend-service: 添加首页配置controller

+++ ruoyi-api/fy-api-goods/src/main/java/com/cloud/mall/fy/api/dto/FirstLevelCategoryDto.java
@@ -15,7 +15,7 @@ public class FirstLevelCategoryDto {
     private Long id;
 
     @ApiModelProperty("当前分类级别")
-    private Byte categoryLevel;
+    private Integer categoryLevel;
 
     @ApiModelProperty("当前一级分类名称")
     private String categoryName;
