+++ ruoyi-modules/fy-modules-goods/src/main/java/com/cloud/mall/fy/goodsservice/entity/GoodsInfo.java
@@ -70,5 +70,5 @@ public class GoodsInfo extends BaseEntityForMall {
     private Integer goodsSellStatus;
 
     @TableField(exist = false)
-    private Byte isDeleted;
+    private Integer isDeleted;
 }
